Add support for chat titles changing
There are times where the title of a group chat is changed, but it's still the same conversation. Rather than storing a single `title` for a MessageFile, there is now a set called `chat_titles`.

To help track this, I've also started adding DB support. There's a single table called "group_chats" with an id field and a name field. All of the entries with the same id are the same conversation, despite the changing names. When we later insert messages into the db, we can use that group_chat.id to link them across title changes. This should allow someone to fetch all messages for a group chat despite the title changing.