Add option to store absolute or relative paths in the DB
References to the messages file or any media files in the DB will default to being relative to the dataset_path. This is the default because 1) it would be desired when distributing the DB and 2) it doesn't reveal any information that could be contained in the full path.

Passing `-a` or `--absolute-paths` will store absolute paths in the DB, which could be helpful if you're working with a copy of the DB that you generated locally.